# Streamline Travis testing