--- a/CIbuild.sh
+++ b/CIbuild.sh
@@ -10,8 +10,8 @@ cmake . -DBUILD_TOOLS=1 -DSELF_TEST=1;
 cd src
 lua CheckBasicStyle.lua
 cd ..
-make -j 2;
-make -j 2 test;
+make -j 4;
+make -j 4 test;
 cd MCServer/;
 if [ "$TRAVIS_MCSERVER_BUILD_TYPE" != "COVERAGE" ]
 	then echo stop | $MCSERVER_PATH;