* Fixes #1795
- Reduced matrix build count from 9 to 3