@worktycho 2015-03-10
I think we need to do a comparison of speeds between 2 and 4 jobs. I would not expect travis vms to have enough parallel resources to run 4 jobs. They only have 1.5 CPU cores.