@bearbin 2015-03-10
I didn't notice that. Yeah 2 would be better, it would reduce the amount the logs get jumbled up.