@tigerw 2015-03-10
I was under the impression that logs only became misaligned with really large thread numbers. Haven't observed any disordering with four threads (on a physical computer) yet.

Isn't the general consensus twice the number of cores available? So maybe three threads. Or four for luck. One can't to too much harm, surely.

Anyway, the build doesn't seem to be starting whatever the concurrency settings are.