@tigerw 2015-03-10
If Travis ever decides to upgrade their OS version to one that comes with GCC 4.8, speeds could be farther improved by using their 'container based infrastructure', which has ~~over 9000~~ 2 cores and 4GB of memory. It doesn't allow su elevation, so at present, it cannot be activated as the apt GCC upgrade commands won't work.