@tigerw 2015-03-10
It inexplicably fails when linking polarssl on GCC. Clang builds *look* faster though.