@tigerw 2015-03-11
Many thanks for the diagnosis, @worktycho.

With 4 threads:
`5299.1  4 min 30 sec  about 7 hours ago  clang TRAVIS_MCSERVER_BUILD_TYPE=RELEASE MCSERVER_PATH=./MCServer Linux` (4m 30s)
`5299.2  4 min 14 sec  about 7 hours ago  clang TRAVIS_MCSERVER_BUILD_TYPE=DEBUG MCSERVER_PATH=./MCServer_debug Linux` (4m 14s)

Without:
`5298.5  5 min 41 sec  about 8 hours ago  clang TRAVIS_MCSERVER_BUILD_TYPE=RELEASE MCSERVER_PATH=./MCServer Linux` (5m 41s)
`5298.6  5 min 17 sec  about 8 hours ago  clang TRAVIS_MCSERVER_BUILD_TYPE=DEBUG MCSERVER_PATH=./MCServer_debug Linux` (5m 17s)

The experimental procedure is somewhat flawed, since the second test is for the NetworkIPEnum branch, but it seems faster.