@worktycho 2015-03-11
My main problem with the experiment is that in values are within normal variations. For example job 5304 has builds lasting between 8 min 22 sec and 3 min 55 sec.