@madmaxoft 2015-03-11
For me the main point is that there are many less CI build jobs, so the overall time is much shorter - ~12 minutes instead of ~40. I'd say merge.