@worktycho 2015-03-11
My view is similar but I think we should remove the change to the number of parallel make jobs then merge. That can be done in a seperate PR.